redid the input because it got deleted last time somehow
somehow I never actually pushed the code that allows for user input.
and then my local copy got deleted.
Also I found a problem in bif.
unbifurcating things, when creatinga new object, wouldn't make it so the
components knew about it, so when repeating it, it would create a new
object each time.
That is fixed now.