"simplify" code for BIFURCATEion, as well as ...
"simplify" code for BIFURCATEion, as well as setting up code for USER
DEFINED FUNCTIONS. OH MY GOSH, USER DEFINED FUNCTIONS ARE IN THERE LOOK
AT HOW THE RETURN IS GOING TO BE.
haha. the joke is that someone would have that reaction.
yeah wnyway thats what this update is about.