Added PRINT2 and importf
importf imports a user defined function from another file.
syntax is
importf filename as FUNCTIONNAME;
PRINT2 prints the contents of a "string", uses syntax:
PRINT2 VARIABLENAME;
yay.
ALSO addition is built in currently, but you can import a better
algorithm for it,
and it probably wont be included to start with later on.
YAY PROGRESS!